member table

+++ routes/web.php
@@ -22,9 +22,9 @@ Route::middleware(['auth', 'admin'])->group(function () {
     Route::get('/members', [MemberController::class, 'index'])->name('admin.member.index');
     Route::get('/members/create', [MemberController::class, 'create'])->name('member.create');
     Route::post('/members', [MemberController::class, 'store'])->name('member.store');
-    Route::get('/members/{id}/edit', [MemberController::class, 'edit'])->name('member.edit');
-    Route::patch('/members/{id}', [MemberController::class, 'update'])->name('member.update');
-    Route::delete('/members/{id}', [MemberController::class, 'destroy'])->name('member.destroy');
+    Route::get('/members/{id}/edit', [MemberController::class, 'edit'])->name('admin.member.edit');
+    Route::patch('/members/{id}', [MemberController::class, 'update'])->name('admin.member.update');
+    Route::delete('/members/{id}', [MemberController::class, 'destroy'])->name('admin.member.destroy');
 });
 
 Route::middleware('auth')->group(function () {
